02_deploy_get_stock_data_upload_storage/requirements.txt
Flask==2.1.0
gunicorn==20.1.0
yfinance
pandas
google-cloud-storage